requirements.txt
streamlit>=1.34
pandas>=2.0
pdfplumber
pypdf
reportlab>=4.0
pymupdf
openpyxl>=3.1.0
streamlit-aggrid 
requests
beautifulsoup4
lxml
pytesseract
pillow
msoffcrypto-tool>=5.0
xlwt




google-api-python-client>=2.0.0
google-auth>=2.0.0
